// apps/web/src/types/generated/PersonalTradeHistoryRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CivilDate } from "./CivilDate";
import type { Side } from "./Side";
import type { StockCode } from "./StockCode";
import type { TradeHistoryReceiptCursor } from "./TradeHistoryReceiptCursor";

export type PersonalTradeHistoryRequest = {
  date_from: CivilDate;
  date_to: CivilDate;
  code: StockCode | null;
  side: Side | null;
  before_receipt: TradeHistoryReceiptCursor | null;
  as_of_receipt: TradeHistoryReceiptCursor | null;
  page_size: number;
};